feat: add Domain, Concept and Question policies with authorization
- Fix SoftDeletes casing and HasMany typo in Concept model
- Add domain_id to concepts migration and fillable
- Register all policies in AppServiceProvider
- Add authorize() calls in controllers

+++ app/Http/Controllers/ConceptController.php
@@ -28,11 +28,15 @@ class ConceptController extends Controller
 
     public function edit(Concept $concept)
     {
+        $this->authorize('update', $concept);
+
         return view('concepts.edit', compact('concept'));
     }
 
     public function update(ConceptRequest $request, Concept $concept)
     {
+        $this->authorize('update', $concept);
+
         $concept->update($request->validated());
 
         return redirect()->route('concepts.index');
@@ -40,6 +44,8 @@ class ConceptController extends Controller
 
     public function destroy(Concept $concept)
     {
+        $this->authorize('delete', $concept);
+
         $concept->delete();
 
         return redirect()->route('concepts.index');
